Fix: Plugin update not hot reloading

+++ src/windows/Store/StoreBackend.py
@@ -620,7 +620,7 @@ class StoreBackend:
         gl.signal_manager.trigger_signal(Signals.PluginInstall, plugin_dict["id"])
 
         log.success(f"Plugin {plugin_dict['id']} installed successfully under: {local_path} with sha: {plugin_dict['commit_sha']}")
-    def uninstall_plugin(self, plugin_id:str, remove_from_pages:bool = False) -> bool:
+    def uninstall_plugin(self, plugin_id:str, remove_from_pages:bool = False, remove_files:bool = True) -> bool:
         ## 1. Remove all action objects in all pages
         for deck_controller in gl.deck_manager.deck_controller:
             # Track all keys controlled by this plugin
@@ -640,10 +640,11 @@ class StoreBackend:
         plugin = gl.plugin_manager.get_plugin_by_id(plugin_id)
         if plugin is None:
             return
-        plugin.on_uninstall()
-        
-        ## 3. Remove plugin folder
-        shutil.rmtree(plugin.PATH)
+        if remove_files:
+            plugin.on_uninstall()
+            
+            ## 3. Remove plugin folder
+            shutil.rmtree(plugin.PATH)
 
         ## 4. Delete plugin base object
         # plugin_obj = gl.plugin_manager.get_plugin_by_id(plugin_id)
@@ -725,6 +726,10 @@ class StoreBackend:
         if isinstance(plugins_to_update, NoConnectionError):
             return plugins_to_update
         for plugin in plugins_to_update:
+            try:
+                await self.uninstall_plugin(plugin["id"], remove_from_pages=False, remove_files=False)
+            except Exception as e:
+                log.error(e)
             await self.install_plugin(plugin)
         
         return len(plugins_to_update)
